@@ -2,11 +2,11 @@ name: venera
 description: "A comic app."
 publish_to: 'none'
 
-version: 1.0.8+108
+version: 1.1.0+110
 
 environment:
-  sdk: '>=3.5.0 <4.0.0'
-  flutter: 3.24.5
+  sdk: '>=3.6.0 <4.0.0'
+  flutter: 3.27.0
 
 dependencies:
   flutter:
@@ -14,7 +14,7 @@ dependencies:
   path_provider: any
   flutter_localizations:
     sdk: flutter
-  intl: any
+  intl: ^0.19.0
   window_manager: ^0.4.3
   sqlite3: ^2.4.7
   sqlite3_flutter_libs: any
@@ -39,7 +39,7 @@ dependencies:
       url: https://github.com/venera-app/flutter.widgets
       ref: 09e756b1f1b04e6298318d99ec20a787fb360f59
       path: packages/scrollable_positioned_list
-  flutter_reorderable_grid_view: 5.3.2
+  flutter_reorderable_grid_view: ^5.4.0
   yaml: any
   uuid: ^4.5.1
   desktop_webview_window:
@@ -58,7 +58,7 @@ dependencies:
     git:
       url: https://github.com/venera-app/lodepng_flutter
       ref: d1c96cd6503103b3270dfe2f320d4a1c93780f53
-  rhttp: 0.9.1
+  rhttp: 0.9.6
   webdav_client:
     git:
       url: https://github.com/wgh136/webdav_client
@@ -70,11 +70,14 @@ dependencies:
       url: https://github.com/pkuislm/flutter_saf.git
       ref: 3315082b9f7055655610e4f6f136b69e48228c05
   pdf: ^3.11.1
+  dynamic_color: ^1.7.0
+  shimmer: ^3.0.0
+  flutter_memory_info: ^0.0.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.0
+  flutter_lints: ^5.0.0
   flutter_to_arch:
     git: https://github.com/wgh136/flutter_to_arch
   flutter_to_debian:
