@@ -76,7 +76,7 @@ class _AppbarState extends State<Appbar> {
     var content = Container(
       decoration: BoxDecoration(
         color: widget.backgroundColor ??
-            context.colorScheme.surface.withOpacity(0.72),
+            context.colorScheme.surface.toOpacity(0.72),
       ),
       height: _kAppBarHeight + context.padding.top,
       child: Row(
@@ -189,20 +189,19 @@ class _MySliverAppBarDelegate extends SliverPersistentHeaderDelegate {
         leading ??
             (Navigator.of(context).canPop()
                 ? Tooltip(
-              message: "Back".tl,
-              child: IconButton(
-                icon: const Icon(Icons.arrow_back),
-                onPressed: () => Navigator.maybePop(context),
-              ),
-            )
+                    message: "Back".tl,
+                    child: IconButton(
+                      icon: const Icon(Icons.arrow_back),
+                      onPressed: () => Navigator.maybePop(context),
+                    ),
+                  )
                 : const SizedBox()),
         const SizedBox(
           width: 16,
         ),
         Expanded(
           child: DefaultTextStyle(
-            style:
-            DefaultTextStyle.of(context).style.copyWith(fontSize: 20),
+            style: DefaultTextStyle.of(context).style.copyWith(fontSize: 20),
             maxLines: 1,
             overflow: TextOverflow.ellipsis,
             child: title,
@@ -215,12 +214,12 @@ class _MySliverAppBarDelegate extends SliverPersistentHeaderDelegate {
       ],
     ).paddingTop(topPadding);
 
-    if(style == AppbarStyle.blur) {
+    if (style == AppbarStyle.blur) {
       return SizedBox.expand(
         child: BlurEffect(
           blur: 15,
           child: Material(
-            color: context.colorScheme.surface.withOpacity(0.72),
+            color: context.colorScheme.surface.toOpacity(0.72),
             elevation: 0,
             borderRadius: BorderRadius.circular(radius),
             child: body,
@@ -298,12 +297,21 @@ class _FilledTabBarState extends State<FilledTabBar> {
     super.dispose();
   }
 
+  PageStorageBucket get bucket => PageStorage.of(context);
+
   @override
   void didChangeDependencies() {
     _controller = widget.controller ?? DefaultTabController.of(context);
-    _controller.animation!.addListener(onTabChanged);
     initPainter();
     super.didChangeDependencies();
+    var prevIndex = bucket.readState(context) as int?;
+    if (prevIndex != null &&
+        prevIndex != _controller.index &&
+        prevIndex >= 0 &&
+        prevIndex < widget.tabs.length) {
+      _controller.index = prevIndex;
+    }
+    _controller.animation!.addListener(onTabChanged);
   }
 
   @override
@@ -347,6 +355,7 @@ class _FilledTabBarState extends State<FilledTabBar> {
       controller: scrollController,
       builder: (context, controller, physics) {
         return SingleChildScrollView(
+          key: const PageStorageKey('scroll'),
           scrollDirection: Axis.horizontal,
           padding: EdgeInsets.zero,
           controller: controller,
@@ -387,6 +396,7 @@ class _FilledTabBarState extends State<FilledTabBar> {
     }
     updateScrollOffset(i);
     previousIndex = i;
+    bucket.writeState(context, i);
   }
 
   void updateScrollOffset(int i) {
@@ -724,6 +734,7 @@ class _SliverSearchBarDelegate extends SliverPersistentHeaderDelegate {
                       icon: const Icon(Icons.clear),
                       onPressed: () {
                         editingController.clear();
+                        onChanged?.call("");
                       },
                     );
             },
