@@ -214,7 +214,7 @@ class _ButtonState extends State<Button> {
             boxShadow: (isHover && !isLoading && (widget.type == ButtonType.filled || widget.type == ButtonType.normal))
                 ? [
                     BoxShadow(
-                      color: Colors.black.withOpacity(0.1),
+                      color: Colors.black.toOpacity(0.1),
                       blurRadius: 2,
                       offset: const Offset(0, 1),
                     )
@@ -248,7 +248,7 @@ class _ButtonState extends State<Button> {
     if (widget.type == ButtonType.filled) {
       var color = widget.color ?? context.colorScheme.primary;
       if (isHover) {
-        return color.withOpacity(0.9);
+        return color.toOpacity(0.9);
       } else {
         return color;
       }
@@ -256,13 +256,13 @@ class _ButtonState extends State<Button> {
     if (widget.type == ButtonType.normal) {
       var color = widget.color ?? context.colorScheme.surfaceContainer;
       if (isHover) {
-        return color.withOpacity(0.9);
+        return color.toOpacity(0.9);
       } else {
         return color;
       }
     }
     if (isHover) {
-      return context.colorScheme.outline.withOpacity(0.2);
+      return context.colorScheme.outline.toOpacity(0.2);
     }
     return Colors.transparent;
   }
@@ -345,7 +345,7 @@ class _IconButtonState extends State<_IconButton> {
                   ? Theme.of(context)
                   .colorScheme
                   .outlineVariant
-                  .withOpacity(0.4)
+                  .toOpacity(0.4)
                   : null,
               borderRadius: BorderRadius.circular((iconSize + 12) / 2),
             ),
