@@ -1,7 +1,8 @@
 part of 'components.dart';
 
 class MouseBackDetector extends StatelessWidget {
-  const MouseBackDetector({super.key, required this.onTapDown, required this.child});
+  const MouseBackDetector(
+      {super.key, required this.onTapDown, required this.child});
 
   final Widget child;
 
@@ -17,6 +18,48 @@ class MouseBackDetector extends StatelessWidget {
       },
       behavior: HitTestBehavior.translucent,
       child: child,
+    );
+  }
+}
+
+class AnimatedTapRegion extends StatefulWidget {
+  const AnimatedTapRegion({
+    super.key,
+    required this.child,
+    required this.onTap,
+    this.borderRadius = 0,
+  });
+
+  final Widget child;
+
+  final void Function() onTap;
+
+  final double borderRadius;
+
+  @override
+  State<AnimatedTapRegion> createState() => _AnimatedTapRegionState();
+}
+
+class _AnimatedTapRegionState extends State<AnimatedTapRegion> {
+  bool isHovered = false;
+
+  @override
+  Widget build(BuildContext context) {
+    return MouseRegion(
+      onEnter: (_) => setState(() => isHovered = true),
+      onExit: (_) => setState(() => isHovered = false),
+      child: GestureDetector(
+        onTap: widget.onTap,
+        child: ClipRRect(
+          borderRadius: BorderRadius.circular(widget.borderRadius),
+          clipBehavior: Clip.antiAlias,
+          child: AnimatedScale(
+            duration: _fastAnimationDuration,
+            scale: isHovered ? 1.1 : 1,
+            child: widget.child,
+          ),
+        ),
+      ),
     );
   }
 }
