@@ -10,7 +10,9 @@ import 'package:venera/foundation/comic_source/comic_source.dart';
 import 'package:venera/foundation/comic_type.dart';
 import 'package:venera/foundation/consts.dart';
 import 'package:venera/foundation/favorites.dart';
+import 'package:venera/foundation/local.dart';
 import 'package:venera/foundation/res.dart';
+import 'package:venera/network/download.dart';
 import 'package:venera/pages/comic_page.dart';
 import 'package:venera/utils/io.dart';
 import 'package:venera/utils/translations.dart';
@@ -92,7 +94,7 @@ class _FavoritesPageState extends State<FavoritesPage>  {
       barrierDismissible: true,
       fullscreenDialog: true,
       opaque: false,
-      barrierColor: Colors.black.withOpacity(0.36),
+      barrierColor: Colors.black.toOpacity(0.36),
       pageBuilder: (context, animation, secondary) {
         return Align(
           alignment: Alignment.centerLeft,
@@ -150,14 +152,14 @@ class _FavoritesPageState extends State<FavoritesPage>  {
       );
     }
     if (!isNetwork) {
-      return _LocalFavoritesPage(folder: folder!, key: Key(folder!));
+      return _LocalFavoritesPage(folder: folder!, key: PageStorageKey(folder!));
     } else {
       var favoriteData = getFavoriteDataOrNull(folder!);
       if (favoriteData == null) {
         folder = null;
         return buildBody();
       } else {
-        return NetworkFavoritePage(favoriteData, key: Key(folder!));
+        return NetworkFavoritePage(favoriteData, key: PageStorageKey(folder!));
       }
     }
   }
